Fixed tests and unified implementation
