+++ simphony/cuds/__init__.py
@@ -1,7 +1,8 @@
 from .abc_lattice import ABCLattice
 from .abc_mesh import ABCMesh
 from .abc_particles import ABCParticles
-from .mesh import Mesh, Point, Element, Edge, Face, Cell
+from .mesh import Mesh
+from .mesh_items import Point, Element, Edge, Face, Cell
 from .lattice import Lattice, LatticeNode
 from .particles import Particles, Particle, Bond
 from .model import CUDS
