+++ simphony/cuds/abc_mesh.py
@@ -1,6 +1,12 @@
+import uuid
+import itertools
+from abc import abstractmethod
+
 from simphony.core.cuds_item import CUDSItem
 from simphony.cuds.abc_dataset import ABCDataset
 from simphony.cuds.utils import deprecated
+
+from simphony.cuds.mesh_items import Point, Edge, Face, Cell
 
 
 class ABCMesh(ABCDataset):
@@ -11,6 +17,136 @@ class ABCMesh(ABCDataset):
     name : str
         name of mesh
     """
+
+    # Implements ABCDataset interface
+    def get(self, uid):
+        """ Returns an edge with a given uid.
+
+        Returns the edge stored in the mesh
+        identified by uid. If such edge do not
+        exists an exception is raised.
+
+        Parameters
+        ----------
+        uid : uuid.UUID
+            uid of the desired edge.
+
+        Returns
+        -------
+        edge : Edge
+            Edge identified by uid
+
+        Raises
+        ------
+        KeyError :
+            If the edge identified by uid was not found
+        TypeError :
+            When ``uid`` is not uuid.UUID
+        """
+
+        if not isinstance(uid, uuid.UUID):
+            message = 'Expected type for `uid` is uuid.UUID but received {!r}'
+            raise TypeError(message.format(type(uid)))
+
+        try:
+            return self._get_point(uid)
+        except KeyError:
+            pass
+
+        try:
+            return self._get_edge(uid)
+        except KeyError:
+            pass
+
+        try:
+            return self._get_face(uid)
+        except KeyError:
+            pass
+
+        try:
+            return self._get_cell(uid)
+        except KeyError:
+            pass
+
+        raise KeyError("Unknown uid {}".format(uid))
+
+    def add(self, items):
+        uids = []
+        for item in items:
+            if isinstance(item, Point):
+                uids.extend(self._add_points([item]))
+            elif isinstance(item, Edge):
+                uids.extend(self._add_edges([item]))
+            elif isinstance(item, Face):
+                uids.extend(self._add_faces([item]))
+            elif isinstance(item, Cell):
+                uids.extend(self._add_cells([item]))
+            else:
+                raise TypeError(
+                    "Unrecognised item type {!r}".format(item)
+                )
+        return uids
+
+    def update(self, items):
+        for item in items:
+            if isinstance(item, Point):
+                self._update_points([item])
+            elif isinstance(item, Edge):
+                self._update_edges([item])
+            elif isinstance(item, Face):
+                self._update_faces([item])
+            elif isinstance(item, Cell):
+                self._update_cells([item])
+            else:
+                raise TypeError(
+                    "Unrecognised item type {!r}".format(item)
+                )
+
+    def remove(self, uids):
+        raise NotImplementedError("Remove is not implemented for Mesh")
+
+    def iter(self, uids=None, item_type=None):
+        if item_type == CUDSItem.POINT:
+            return self._iter_points(uids)
+        elif item_type == CUDSItem.EDGE:
+            return self._iter_edges(uids)
+        elif item_type == CUDSItem.FACE:
+            return self._iter_faces(uids)
+        elif item_type == CUDSItem.CELL:
+            return self._iter_cells(uids)
+        else:
+            if uids is None:
+                return itertools.chain(
+                    self._iter_points(),
+                    self._iter_edges(),
+                    self._iter_faces(),
+                    self._iter_cells(),
+                )
+            else:
+                return self._iter_uids(uids)
+
+    def has(self, uid):
+        try:
+            self.get(uid)
+        except KeyError:
+            return False
+
+        return True
+
+    def has_type(self, item_type):
+        if item_type == CUDSItem.POINT:
+            return self._has_points()
+        elif item_type == CUDSItem.EDGE:
+            return self._has_edges()
+        elif item_type == CUDSItem.FACE:
+            return self._has_faces()
+        elif item_type == CUDSItem.CELL:
+            return self._has_cells()
+        else:
+            raise ValueError("Unknown item_type "
+                             "{}".format(item_type))
+
+    # Deprecated methods.
 
     @deprecated
     def get_point(self, uid):  # pragma: no cover
@@ -437,3 +573,85 @@ class ABCMesh(ABCDataset):
             False otherwise
         """
         return self.has_type(CUDSItem.CELL)
+
+    # Private. Need to be reimplemented by subclasses
+
+    @abstractmethod
+    def _get_point(self, uid):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _get_edge(self, uid):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _get_face(self, uid):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _get_cell(self, uid):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _add_points(self, points):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _add_edges(self, edges):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _add_faces(self, faces):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _add_cells(self, cells):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _update_points(self, points):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _update_edges(self, edges):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _update_faces(self, faces):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _update_cells(self, cells):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _iter_points(self, uids=None):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _iter_edges(self, uids=None):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _iter_faces(self, uids=None):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _iter_cells(self, uids=None):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _has_points(self):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _has_edges(self):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _has_faces(self):  # pragma: no cover
+        pass
+
+    @abstractmethod
+    def _has_cells(self):  # pragma: no cover
+        pass
