+++ simphony/cuds/mesh.py
@@ -5,165 +5,11 @@ and modify a mesh
 
 """
 import uuid
-import itertools
 
-from simphony.cuds.abc_mesh import ABCMesh
-from simphony.core.cuds_item import CUDSItem
 import simphony.core.data_container as dc
-
-
-class Point(object):
-    """ Coordinates describing a point in the space
-
-    Set of coordinates (x,y,z) describing a point in
-    the space and data about that point
-
-    Parameters
-    ----------
-    uid : uuid.UUID
-        uid of the point.
-    coordinates : list of double
-        set of coordinates (x,y,z) describing the point position.
-    data : DataContainer
-        object to store point data
-
-    Attributes
-    ----------
-    uid : uuid.UUID
-        uid of the point.
-    data : DataContainer
-        object to store point data
-    coordinates : list of double
-        set of coordinates (x,y,z) describing the point position.
-
-    """
-
-    def __init__(self, coordinates, uid=None, data=None):
-        self.uid = uid
-        self.coordinates = tuple(coordinates)
-
-        if data:
-            self.data = dc.DataContainer(data)
-        else:
-            self.data = dc.DataContainer()
-
-    @classmethod
-    def from_point(cls, point):
-        return cls(
-            point.coordinates,
-            point.uid,
-            point.data
-        )
-
-
-class Element(object):
-    """ Element base class
-
-    Element for storing geometrical objects
-
-    Parameters
-    ----------
-    uid :
-        uid of the edge.
-    points : list of uid
-        list of points uids defining the edge.
-    data : DataContainer
-        object to store data relative to the element
-
-    Attributes
-    ----------
-    points : list of uid
-        list of points uids defining the element.
-    uid : uuid.UUID
-        uid of the element
-    data : DataContainer
-        Element data
-
-    """
-
-    def __init__(self, points, uid=None, data=None):
-        self.uid = uid
-        self.points = points[:]
-
-        if data:
-            self.data = dc.DataContainer(data)
-        else:
-            self.data = dc.DataContainer()
-
-
-class Edge(Element):
-    """ Edge element
-
-    Element for storing 1D geometrical objects
-
-    Parameters
-    ----------
-    points : list of uid
-        list of points uids defining the edge.
-    uid : uuid.UUID
-        uid of the edge.
-    data : DataContainer
-        object to store data relative to the edge
-
-    """
-
-    @classmethod
-    def from_edge(cls, edge):
-        return cls(
-            edge.points,
-            edge.uid,
-            edge.data
-        )
-
-
-class Face(Element):
-    """ Face element
-
-    Element for storing 2D geometrical objects
-
-    Parameters
-    ----------
-    points : list of uid
-        list of points uids defining the face.
-    uid : uuid.UUID
-        uid of the face.
-    data : DataContainer
-        object to store data relative to the face
-
-    """
-
-    @classmethod
-    def from_face(cls, face):
-        return cls(
-            face.points,
-            face.uid,
-            face.data
-        )
-
-
-class Cell(Element):
-    """ Cell element
-
-    Element for storing 3D geometrical objects
-
-    Parameters
-    ----------
-    points : list of uid
-        list of points uids defining the cell.
-    uid : uuid.UUID
-        uid of the cell.
-    data : DataContainer
-        object to store data relative to the cell
-
-    """
-
-    @classmethod
-    def from_cell(cls, cell):
-        return cls(
-            cell.points,
-            cell.uid,
-            cell.data
-        )
+from simphony.core.cuds_item import CUDSItem
+from simphony.cuds.abc_mesh import ABCMesh
+from simphony.cuds.mesh_items import Edge, Face, Cell, Point
 
 
 class Mesh(ABCMesh):
@@ -253,129 +99,6 @@ class Mesh(ABCMesh):
         except KeyError:
             error_str = "Trying to obtain count a of non-supported item: {}"
             raise ValueError(error_str.format(item_type))
-
-    def get(self, uid):
-        """ Returns an item with a given uid.
-
-        Returns the item stored in the mesh
-        identified by uid. If such item does not
-        exists an exception is raised.
-
-        Parameters
-        ----------
-        uid : uuid.UUID
-            uid of the desired item.
-
-        Returns
-        -------
-        item : Point | Edge | Face | Cell
-            Item identified by uid
-
-        Raises
-        ------
-        KeyError :
-            If the item identified by uid was not found
-        TypeError :
-            When ``uid`` is not uuid.UUID
-        """
-
-        try:
-            return self._get_point(uid)
-        except KeyError:
-            pass
-
-        try:
-            return self._get_edge(uid)
-        except KeyError:
-            pass
-
-        try:
-            return self._get_face(uid)
-        except KeyError:
-            pass
-
-        try:
-            return self._get_cell(uid)
-        except KeyError:
-            pass
-
-        raise KeyError("Unknown uid {}".format(uid))
-
-    def add(self, items):
-        uids = []
-        for item in items:
-            if isinstance(item, Point):
-                uids.extend(self._add_points([item]))
-            elif isinstance(item, Edge):
-                uids.extend(self._add_edges([item]))
-            elif isinstance(item, Face):
-                uids.extend(self._add_faces([item]))
-            elif isinstance(item, Cell):
-                uids.extend(self._add_cells([item]))
-            else:
-                raise TypeError(
-                    "Unrecognised item type {!r}".format(item)
-                )
-        return uids
-
-    def update(self, items):
-        for item in items:
-            if isinstance(item, Point):
-                self._update_points([item])
-            elif isinstance(item, Edge):
-                self._update_edges([item])
-            elif isinstance(item, Face):
-                self._update_faces([item])
-            elif isinstance(item, Cell):
-                self._update_cells([item])
-            else:
-                raise TypeError(
-                    "Unrecognised item type {!r}".format(item)
-                )
-
-    def remove(self, uids):
-        raise NotImplementedError("Remove is not implemented for Mesh")
-
-    def iter(self, uids=None, item_type=None):
-        if item_type == CUDSItem.POINT:
-            return self._iter_points(uids)
-        elif item_type == CUDSItem.EDGE:
-            return self._iter_edges(uids)
-        elif item_type == CUDSItem.FACE:
-            return self._iter_faces(uids)
-        elif item_type == CUDSItem.CELL:
-            return self._iter_cells(uids)
-        else:
-            if uids is None:
-                return itertools.chain(
-                    self._iter_points(),
-                    self._iter_edges(),
-                    self._iter_faces(),
-                    self._iter_cells(),
-                )
-            else:
-                return self._iter_uids(uids)
-
-    def has(self, uid):
-        try:
-            self.get(uid)
-        except KeyError:
-            return False
-
-        return True
-
-    def has_type(self, item_type):
-        if item_type == CUDSItem.POINT:
-            return self._has_points()
-        elif item_type == CUDSItem.EDGE:
-            return self._has_edges()
-        elif item_type == CUDSItem.FACE:
-            return self._has_faces()
-        elif item_type == CUDSItem.CELL:
-            return self._has_cells()
-        else:
-            raise ValueError("Unknown item_type "
-                             "{}".format(item_type))
 
     # Private
