+++ simphony/testing/abc_check_mesh.py
@@ -6,7 +6,7 @@ from functools import partial
 from simphony.testing.utils import (
     create_data_container, create_points, compare_points, compare_elements,
     grouper, compare_data_containers)
-from simphony.cuds.mesh import Point, Edge, Cell, Face
+from simphony.cuds.mesh_items import Edge, Face, Cell, Point
 from simphony.core.cuds_item import CUDSItem
 from simphony.core.data_container import DataContainer
